v4.19 Changed / fixed some things and added a few features.
* Changed the minimum width of popups.
* Fixed a bug that was popping out popouts inside of their boxes,
causing them to overflow.
* Fixed an issue that was scrolling the page a bit upwards if a
referenced comment was at the end of the page.
* Comments referenced after the page has been loaded are now highlighted
with a green arrow for easier spotting.
* Fixed a bug where the remove links from the entered / whitelist /
blacklist pages were not working after page 2 in Endless Scrolling
(closes #47).
* Removed the functionality to notify if a gift has been delivered from
Header Icons Refresher and added it as a separate feature: Delivered
Gifts Notifier.
* Changed the color of the icon in Delivered Gifts Notifier.
* Fixed a bug that was notifying a delivered gift if it was marked as
not received in Delivered Gifts Notifier.
* Removed the option to move active discussions to the top of the page
from Endless Scrolling and added it as a separate feature: Active
Discussions On Top.
* Rows are now highlighted upon hovering over them in User Giveaways
Data (closes #55).
* Cells with a value of 0 are now faded out in User Giveaways Data
(closes #56).
* Fixed a bug that was trying to enable Featured Container Hider on
SteamTrades and therefore preventing the script from loading.
* Added Groups Status (closes #40).